data_agent/requirements.txt
langchain_community
langchain_groq
langchain_experimental
wikipedia
tabulate
openpyxl
